+++ CHANGELOG.md
@@ -126,3 +126,9 @@ All notable changes to `redis-helpers` will be documented in this file
 ## 1.3.5 - 2021-08-26
 - cut laravel/framework dependency & replaced with illuminate/support
 - add `NestedKeysTest` to unit test traits so that it can be tested using 'mock' & 'predis' clients
+
+
+## 1.4.0 - 2021-08-30
+- optimize test suite organization by setting redis client & other config values in the root `TestCase` #26
+- fix issues with nested keys tests failing while using the 'mock' redis client #27
+- add try/catch block to `RedisCache::key()` method so that when an error is encountered while trying to find child keys, the original key is returned